# app/api/projects/route.ts
import { prisma } from '@/lib/prisma';
import { broadcastAdminUpdate } from '@/lib/sse/sse-service';
import { NextResponse } from 'next/server';
import { Prisma, Character } from '@prisma/client';
import { auth } from '@/lib/auth';
import { createProjectSchema } from '@/lib/schemas';

export const dynamic = 'force-dynamic';

export async function POST(request: Request) {
    try {
        const session = await auth();
        if (!session?.user?.id) {
            return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
        }

        const body = await request.json();

        // Validate input using Zod
        const validation = createProjectSchema.safeParse(body);
        if (!validation.success) {
            return NextResponse.json({
                error: 'Validation Error',
                details: validation.error.format()
            }, { status: 400 });
        }

        const {
            topic,
            style,
            language,
            voice_name,
            tts_provider,
            reference_image_url,
            characterIds,
            include_music,
            bg_music_prompt,
            duration_config,
            folder_id,
            generated_title,
            generated_description,
            generated_shorts_hashtags,
            generated_tiktok_text,
            generated_tiktok_hashtags,
            script_metadata,
            channel_id, // 🆕 Canal de destino
            persona_id, // 🆕 Persona usada
        } = validation.data;

        const user_id = session.user.id;

        // 🆕 Se channelId fornecido, buscar persona do canal
        let effectivePersonaId = persona_id;
        let personaVersion = null;

        if (channel_id && !persona_id) {
            const channel = await prisma.channel.findUnique({
                where: { id: channel_id },
                include: { persona: true }
            });

            if (channel?.personaId) {
                effectivePersonaId = channel.personaId;
                personaVersion = channel.persona?.version;
            }
        } else if (persona_id) {
            const persona = await prisma.persona.findUnique({
                where: { id: persona_id }
            });
            personaVersion = persona?.version;
        }

        // --- LIMIT CHECK ---
        const user = await prisma.user.findUnique({
            where: { id: user_id },
            include: { user_limits: true }
        });

        if (!user) return NextResponse.json({ error: 'User not found' }, { status: 404 });

        let limits = user.user_limits;
        if (!limits) {
            limits = await prisma.userLimits.create({ data: { user_id } });
        }

        const today = new Date();
        const lastReset = new Date(limits.last_daily_reset);
        const isNewDay = today.getDate() !== lastReset.getDate() || today.getMonth() !== lastReset.getMonth();

        if (isNewDay) {
            // Reset daily counters
            limits = await prisma.userLimits.update({
                where: { user_id },
                data: {
                    current_daily_requests: 0,
                    current_daily_videos: 0,
                    last_daily_reset: today
                }
            });
        }

        // Check Limit for FREE users
        const isFree = user.subscription_plan === 'FREE' || user.tier === 'free';
        if (isFree) {
            const limit = limits.daily_videos_limit; // Default is 1 in schema
            if (limits.current_daily_videos >= limit) {
                return NextResponse.json({
                    error: 'Daily video limit reached',
                    details: 'Free accounts are limited to 1 video per day. Upgrade to Pro for more.'
                }, { status: 403 });
            }
        }

        // Increment Video Count
        await prisma.userLimits.update({
            where: { user_id },
            data: {
                current_daily_videos: { increment: 1 },
                current_videos_used: { increment: 1 }
            }
        });
        // -------------------

        // If characterIds provided, fetch character data and create snapshot
        let reference_characters_snapshot: Prisma.InputJsonValue | undefined;
        if (characterIds && Array.isArray(characterIds) && characterIds.length > 0) {
            const characters = await prisma.character.findMany({
                where: {
                    id: { in: characterIds },
                    user_id, // Ensure characters belong to the user
                },
            });

            // Create snapshot in the format expected by frontend
            reference_characters_snapshot = characters.map((char: Character) => ({
                id: char.id,
                name: char.name,
                description: char.description,
                images: char.images,
            })) as Prisma.InputJsonValue;
        }

        const project = await prisma.project.create({
            data: {
                user_id,
                topic,
                style,
                language,
                voice_name,
                tts_provider,
                reference_image_url,
                // We still support snapshot for legacy/backup, but primary link is relation
                ...(reference_characters_snapshot && { reference_characters_snapshot }),
                include_music,
                bg_music_prompt,
                bg_music_status: include_music ? 'pending' : null,
                duration_config: duration_config || Prisma.JsonNull,
                status: 'draft',
                folder_id,
                generated_title,
                generated_description,
                generated_shorts_hashtags: generated_shorts_hashtags || [],
                generated_tiktok_text,
                generated_tiktok_hashtags: generated_tiktok_hashtags || [],
                script_metadata: script_metadata || Prisma.JsonNull,
                // 🆕 Novos campos de Personas & Channels
                channel_id: channel_id || null,
                persona_id: effectivePersonaId || null,
                persona_version: personaVersion,
                // Explicitly create relation records
                ProjectCharacters: {
                    create: characterIds?.map((id: string) => ({
                        characters: { connect: { id } }
                    })) || []
                }
            },
            include: {
                ProjectCharacters: {
                    include: { characters: true }
                },
                channel: true,
                persona: {
                    select: {
                        id: true,
                        name: true,
                        type: true,
                        category: true
                    }
                }
            },
        });

        // Map back to expected structure
        const mappedProject = {
            ...project,
            ProjectCharacters: undefined, // Remove the join table prop
            characters: project.ProjectCharacters.map(pc => pc.characters)
        };

        // Broadcast to admin dashboard
        broadcastAdminUpdate('PROJECT_CREATED', mappedProject);

        return NextResponse.json(mappedProject);
    } catch (error: any) {
        console.error('Error creating project:', error);
        return NextResponse.json({ error: 'Internal Server Error' }, { status: 500 });
    }
}

export async function GET(request: Request) {
    try {
        const session = await auth();
        if (!session?.user?.id) {
            return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
        }

        const { searchParams } = new URL(request.url);
        const user_id = session.user.id;
        const limitParam = searchParams.get('limit');
        const offsetParam = searchParams.get('offset');

        const limit = limitParam ? parseInt(limitParam) : undefined;
        const offset = offsetParam ? parseInt(offsetParam) : undefined;

        const folderIdParam = searchParams.get('folder_id');
        const isArchivedParam = searchParams.get('is_archived');

        const where: any = { user_id };

        if (folderIdParam === 'root') {
            where.folder_id = null;
        } else if (folderIdParam) {
            where.folder_id = folderIdParam;
        }

        if (isArchivedParam !== null) {
            where.is_archived = isArchivedParam === 'true';
        }

        const [projects, total] = await Promise.all([
            prisma.project.findMany({
                where,
                orderBy: { created_at: 'desc' },
                take: limit,
                skip: offset,
                select: {
                    id: true,
                    user_id: true,
                    topic: true,
                    style: true,
                    language: true,
                    voice_name: true,
                    tts_provider: true,
                    created_at: true,
                    generated_title: true,
                    generated_description: true,
                    generated_shorts_hashtags: true,
                    generated_tiktok_text: true,
                    generated_tiktok_hashtags: true,
                    script_metadata: true,
                    status: true,
                    is_archived: true,
                    tags: true,
                    folder_id: true,
                    // Include simplified scenes state
                    scenes: {
                        where: { deleted_at: null },
                        select: {
                            id: true,
                            scene_number: true,
                            image_status: true,
                            audio_status: true,
                            video_status: true,
                            media_type: true,
                            video_model: true
                        },
                        orderBy: { scene_number: 'asc' }
                    },
                    // Include Characters via Explicit Table
                    ProjectCharacters: {
                        select: {
                            characters: {
                                select: {
                                    id: true,
                                    name: true,
                                    images: true,
                                    description: true
                                }
                            }
                        }
                    }
                }
            }),
            prisma.project.count({ where })
        ]);

        // Transform Results to flatten characters
        const mappedProjects = projects.map(p => ({
            ...p,
            ProjectCharacters: undefined,
            characters: p.ProjectCharacters.map(pc => pc.characters)
        }));

        if (limit !== undefined) {
            return NextResponse.json({
                data: mappedProjects,
                meta: {
                    total,
                    limit,
                    offset: offset || 0
                }
            });
        }

        return NextResponse.json(mappedProjects);
    } catch (error: any) {
        console.error('Error fetching projects:', error);
        return NextResponse.json({ error: 'Internal Server Error' }, { status: 500 });
    }
}